backend/scripts/requirements.txt
# PDF Generation
weasyprint==60.1
# AWS S3
boto3==1.34.0
# Database
psycopg2-binary==2.9.9
# Environment Variables
python-dotenv==1.0.0
# HTML Templating
jinja2==3.1.2
# Date handling
python-dateutil==2.8.2
